feat: init login post request and get token back
